pyproject.toml
[project]
name = "tgx"
version = "0.1.0"
description = "Библиотека общих функций и классов для Telegram-ботов"
authors = [{ name = "Stanislav Shavlinsky", email = "stanislave777@gmail.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot>=20.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
